Add systemLog API
This includes a minor SDK refactor as it removes quite a few specialized
functions and facilities from core.h into system.h. There was a breaking point
for needing to frequently update core includes.

The new logger systemLog function allows a callsite to bypass logging config
and write a line to the OS logger (aka syslog).
