+++ osquery/core/system.cpp
@@ -41,6 +41,7 @@
 #include <osquery/filesystem.h>
 #include <osquery/logger.h>
 #include <osquery/sql.h>
+#include <osquery/system.h>
 
 namespace fs = boost::filesystem;
