+++ osquery/core/tests/conversions_tests.cpp
@@ -15,6 +15,7 @@
 
 #include <osquery/core.h>
 #include <osquery/flags.h>
+#include <osquery/system.h>
 
 #include "osquery/core/conversions.h"
