+++ osquery/core/tests/permissions_tests.cpp
@@ -20,6 +20,7 @@
 
 #include <osquery/core.h>
 #include <osquery/logger.h>
+#include <osquery/system.h>
 
 #include "osquery/tests/test_util.h"
