+++ osquery/dispatcher/tests/scheduler_tests.cpp
@@ -11,6 +11,7 @@
 #include <gtest/gtest.h>
 
 #include <osquery/logger.h>
+#include <osquery/system.h>
 
 #include "osquery/tests/test_util.h"
 #include "osquery/sql/sqlite_util.h"
