+++ osquery/events/events.cpp
@@ -21,6 +21,7 @@
 #include <osquery/events.h>
 #include <osquery/flags.h>
 #include <osquery/logger.h>
+#include <osquery/system.h>
 
 #include "osquery/core/conversions.h"
