+++ osquery/events/linux/inotify.cpp
@@ -17,6 +17,7 @@
 
 #include <osquery/filesystem.h>
 #include <osquery/logger.h>
+#include <osquery/system.h>
 
 #include "osquery/events/linux/inotify.h"
