+++ osquery/extensions/interface.cpp
@@ -11,6 +11,7 @@
 #include <osquery/core.h>
 #include <osquery/filesystem.h>
 #include <osquery/logger.h>
+#include <osquery/system.h>
 
 #include "osquery/extensions/interface.h"
