+++ osquery/filesystem/filesystem.cpp
@@ -28,6 +28,7 @@
 #include <osquery/filesystem.h>
 #include <osquery/logger.h>
 #include <osquery/sql.h>
+#include <osquery/system.h>
 
 #include "osquery/filesystem/fileops.h"
