+++ osquery/logger/plugins/buffered.cpp
@@ -18,6 +18,7 @@
 #include <osquery/enroll.h>
 #include <osquery/flags.h>
 #include <osquery/registry.h>
+#include <osquery/system.h>
 
 #include "osquery/config/parsers/decorators.h"
 #include "osquery/logger/plugins/buffered.h"
