+++ osquery/main/shell.cpp
@@ -17,11 +17,12 @@
 #include <iostream>
 
 #include <osquery/core.h>
-#include <osquery/core/process.h>
 #include <osquery/database.h>
 #include <osquery/extensions.h>
 #include <osquery/flags.h>
+#include <osquery/system.h>
 
+#include "osquery/core/process.h"
 #include "osquery/core/watcher.h"
 #include "osquery/devtools/devtools.h"
 #include "osquery/filesystem/fileops.h"
