+++ osquery/tables/events/linux/socket_events.cpp
@@ -12,6 +12,7 @@
 #include <boost/algorithm/string.hpp>
 
 #include <osquery/sql.h>
+#include <osquery/system.h>
 
 #include "osquery/core/conversions.h"
 #include "osquery/events/linux/audit.h"
