+++ osquery/tables/events/linux/user_events.cpp
@@ -11,6 +11,7 @@
 #include <osquery/config.h>
 #include <osquery/logger.h>
 #include <osquery/sql.h>
+#include <osquery/system.h>
 
 #include "osquery/events/linux/audit.h"
