+++ osquery/tables/system/darwin/system_info.cpp
@@ -16,6 +16,7 @@
 #include <boost/algorithm/string.hpp>
 
 #include <osquery/sql.h>
+#include <osquery/system.h>
 #include <osquery/tables.h>
 
 #include "osquery/core/conversions.h"
